Don't hard-code the default constructor assumption
Don't hard-code the assumption that the implementation has a
default constructor.

Eventually it'd be nice if the `ComponentRegistrations` is a really
simple implementation where you get compile-time help when you forget
to register something, but if you want anything special you can
duplicate it yourself (i.e. to selective enable components or pass
additional parameters via the providers).